+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## `master`
+
+- Add `className` option.
+
 ## `v0.3.2` - Sep 16, 2014
 
 - Add `IETextarea` adapter which supports IE8